Rename package to usermanagement
